+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 ## [6.4.0-pre.10] - Unreleased
+### Added
+- Audio.initInstrument() to init instrument.
+
 ### Changed
+- Audio.addInstrument() accepts url to samples json.
 - Use Synthesizer as fallback if instrument not found.
 
 ## [6.4.0-pre.9] - 2026-03-07